support SKT from database
