+++ GenevaIO/src/main/java/org/genevaers/genevaio/dbreader/DBViewColumnSourceReader.java
@@ -66,6 +66,8 @@ public class DBViewColumnSourceReader extends DBReaderBase{
             //We now need to support sort key titles
             //Need to update the generation to include these here
             vcs.setLogicText(rs.getString("EXTRACTCALCLOGIC"));
+            vcs.setSortTitleFieldId(rs.getInt("SORTTITLELRFIELDID"));
+            vcs.setSortTitleLookupId(rs.getInt("SORTTITLELOOKUPID"));
             view.addViewColumnSource(vcs);
         } else {
            //Can there ever be an else here... ?
